requirements_app.txt
aiogram>=3.22.0
pyyaml>=6.0.0
deep-translator>=1.9.2
omegaconf>=2.3.0
llama-cpp-python
num2words
transformers
torch
torchvision
torchaudio
chatterbox-tts
torchcodec
tokenizers
Pillow>=10.0.0
backoff>=2.2.1
langchain>=0.0.286
requests>=2.31.0
urllib3>=2.0.4
python-dotenv>=1.0.0
pydantic>=1.10.13
backoff
aiogram
extensions
configparser
deep_translator
num2words
torchcodec